% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_COAD_Betas}
\alias{TCGA_COAD_Betas}
\title{TCGA COAD Methylation data}
\format{
\subsection{\code{TCGA_COAD_Betas}}{

A data frame with 331316 rows and 223 columns
}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/pancanatlas}
}
\usage{
TCGA_COAD_Betas
}
\description{
A subset of Methylation data from the TCGA PanCancer Atlas Publication for
use as example data for Joint NMF analysis
}
\keyword{datasets}
