% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{TCGA_COAD_vst_RNAseq}
\alias{TCGA_COAD_vst_RNAseq}
\title{TCGA COAD RNAseq data}
\format{
\subsection{\code{TCGA_COAD_vst_RNAseq}}{

A data frame with 16315 rows and 200 columns
}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/pancanatlas}
}
\usage{
TCGA_COAD_vst_RNAseq
}
\description{
A subset of RNAseq data from the TCGA PanCancer Atlas Publication for
use as example data for Joint NMF analysis. Vst normalized using DESeq2.
}
\keyword{datasets}
